Don't crash consumers
Consumers _will no longer crash_ when an exception is raised. Instead,
consumers will move the message that caused the exception to a dead-letter
exchange. Railway will configure the dead-letter exchange automatically.

All messages will be sent to a single dead-letter exchange called
'ipc:errors'. This makes Datadog monitoring easier since we only need to
monitor a single queue (we couldn't figure out a way to have Datadog
query using wildcards).

We still support overriding the name of the dead-letter exchange,
however, applications that do that are then on the hook for setting up
Datadog monitoring for the custom dead-letter exchange.