Consumer logic for deciding which decoder to use
Consumers now check the message metadata for a message encoding type
which is used to decode the message. If one cannot be found, it falls
back to assuming the message is a Base64 encoded binary protobuf. This
allows us to support other formats in the future (i.e. JSON) without
breaking existing message contracts.

This does introduce a public API breaking change, however. The
`Consumer#work` method was renamed to `Consumer#work_with_params` in
order to get access to the message metadata. Any clients that override
the `Consumer#work` method will now need to override the
`Consumer#work_with_params` method instead.